Parsing code into separate chunks to avoid clutter
